fix max fee handling for pulsechain
