+++ app/tray/Account/Requests/TransactionRequest/AdjustFee/index.js
@@ -4,6 +4,7 @@ import BigNumber from 'bignumber.js'
 
 import link from '../../../../../../resources/link'
 import { usesBaseFee } from '../../../../../../resources/domain/transaction'
+import { getMaxTotalFee } from '../../../../../../resources/gas'
 
 const numberFormat = { groupSeparator: '', decimalSeparator: '.' }
 
@@ -27,8 +28,8 @@ function bnToHex(bn) {
   return `0x${bn.toString(16)}`
 }
 
-function limitRange(bn, min = 0, max = 9999e9) {
-  if (bn.gt(max)) return BigNumber(max)
+function limitRange(bn, min = 0, max) {
+  if (max && bn.gt(max)) return BigNumber(max)
   if (bn.lt(min)) return BigNumber(min)
   return bn
 }
@@ -40,27 +41,10 @@ function formatForInput(num, decimals, useWei = false) {
   return useWei ? toDisplayFromWei(BigNumber(num)) : toDisplayFromGwei(BigNumber(num))
 }
 
-function getMaxTotalFee(tx = { chainId: '' }) {
-  const chainId = parseInt(tx.chainId)
-
-  // for ETH-based chains, the max fee should be 2 ETH
-  if ([1, 3, 4, 5, 6, 10, 42, 61, 62, 63, 69, 42161, 421611].includes(chainId)) {
-    return 2 * 1e18
-  }
-
-  // for Fantom, the max fee should be 250 FTM
-  if ([250, 4002].includes(chainId)) {
-    return 250 * 1e18
-  }
-
-  // for all other chains, default to 50 of the chain's currency
-  return 50 * 1e18
-}
-
 const totalFee = ({ gasPrice, baseFee, priorityFee, gasLimit }) =>
   gasPrice ? gasPrice.times(gasLimit) : baseFee.plus(priorityFee).times(gasLimit)
 
-const limitGasUnits = (bn) => limitRange(bn, 0, 12.5e6)
+const limitGasUnits = (bn) => limitRange(bn, 0)
 
 let submitTimeout = null
