+++ main/accounts/index.ts
@@ -9,7 +9,7 @@ import store from '../store'
 import FrameAccount from './Account'
 import ExternalDataScanner, { DataScanner } from '../externalData'
 import Signer from '../signers/Signer'
-import { signerCompatibility as transactionCompatibility, maxFee, SignerCompatibility } from '../transaction'
+import { signerCompatibility as transactionCompatibility, SignerCompatibility } from '../transaction'
 
 import { weiIntToEthInt, hexToInt } from '../../resources/utils'
 import { accountPanelCrumb, signerPanelCrumb } from '../../resources/domain/nav'
@@ -32,6 +32,7 @@ import { ActionType } from '../transaction/actions'
 import { openBlockExplorer } from '../windows/window'
 import { ApprovalType } from '../../resources/constants'
 import { accountNS } from '../../resources/domain/account'
+import { getMaxTotalFee } from '../../resources/gas'
 
 import type { Chain } from '../chains'
 import type { Account, AccountMetadata, Gas } from '../store/state'
@@ -925,10 +926,10 @@ export class Accounts extends EventEmitter {
     return !fee || isNaN(parseInt(fee, 16)) || parseInt(fee, 16) < 0
   }
 
-  private limitedHexValue(hexValue: string, min: number, max: number) {
+  private limitedHexValue(hexValue: string, min: number, max?: number) {
     const value = parseInt(hexValue, 16)
     if (value < min) return intToHex(min)
-    if (value > max) return intToHex(max)
+    if (max && value > max) return intToHex(max)
     return hexValue
   }
 
@@ -1007,7 +1008,7 @@ export class Accounts extends EventEmitter {
     )
 
     // New value
-    const newBaseFee = parseInt(this.limitedHexValue(baseFee, 0, 9999 * 1e9), 16)
+    const newBaseFee = parseInt(this.limitedHexValue(baseFee, 0), 16)
 
     // No change
     if (newBaseFee === currentBaseFee) return
@@ -1017,7 +1018,7 @@ export class Accounts extends EventEmitter {
 
     // New max fee per gas
     const newMaxFeePerGas = newBaseFee + maxPriorityFeePerGas
-    const maxTotalFee = maxFee(tx)
+    const maxTotalFee = getMaxTotalFee(tx)
 
     // Limit max fee
     if (newMaxFeePerGas * gasLimit > maxTotalFee) {
@@ -1044,7 +1045,7 @@ export class Accounts extends EventEmitter {
     )
 
     // New values
-    const newMaxPriorityFeePerGas = parseInt(this.limitedHexValue(priorityFee, 0, 9999 * 1e9), 16)
+    const newMaxPriorityFeePerGas = parseInt(this.limitedHexValue(priorityFee, 0), 16)
 
     // No change
     if (newMaxPriorityFeePerGas === maxPriorityFeePerGas) return
@@ -1053,7 +1054,7 @@ export class Accounts extends EventEmitter {
 
     // New max fee per gas
     const newMaxFeePerGas = currentBaseFee + newMaxPriorityFeePerGas
-    const maxTotalFee = maxFee(tx)
+    const maxTotalFee = getMaxTotalFee(tx)
 
     // Limit max fee
     if (newMaxFeePerGas * gasLimit > maxTotalFee) {
@@ -1080,14 +1081,14 @@ export class Accounts extends EventEmitter {
     const { currentAccount, gasLimit, gasPrice, txType } = this.txFeeUpdate(price, handlerId, userUpdate)
 
     // New values
-    const newGasPrice = parseInt(this.limitedHexValue(price, 0, 9999 * 1e9), 16)
+    const newGasPrice = parseInt(this.limitedHexValue(price, 0), 16)
 
     // No change
     if (newGasPrice === gasPrice) return
 
     const txRequest = this.getTransactionRequest(currentAccount, handlerId)
     const tx = txRequest.data
-    const maxTotalFee = maxFee(tx)
+    const maxTotalFee = getMaxTotalFee(tx)
 
     // Limit max fee
     if (newGasPrice * gasLimit > maxTotalFee) {
@@ -1109,11 +1110,11 @@ export class Accounts extends EventEmitter {
     const { currentAccount, maxFeePerGas, gasPrice, txType } = this.txFeeUpdate(limit, handlerId, userUpdate)
 
     // New values
-    const newGasLimit = parseInt(this.limitedHexValue(limit, 0, 12.5e6), 16)
+    const newGasLimit = parseInt(this.limitedHexValue(limit, 0), 16)
 
     const txRequest = this.getTransactionRequest(currentAccount, handlerId)
     const tx = txRequest.data
-    const maxTotalFee = maxFee(tx)
+    const maxTotalFee = getMaxTotalFee(tx)
 
     const fee = txType === '0x2' ? maxFeePerGas : gasPrice
     if (newGasLimit * fee > maxTotalFee) {
